+++ src/main/java/users/Admin.java
@@ -14,14 +14,8 @@ public abstract class Admin extends User
 {
     public void AddAdmin(String filename)
     {
-        File file = new File("/home/cybereagle3-1/AndroidStudioProjects/FirstApp/midsem_project/src/main/java/filestuff/" + filename);
+        File file = new File("/home/cybereagle3-1/IdeaProjects/midsem-project-csd-207/src/main/java/filestuff/" + filename);
         new FileIO().WriteUsersonFile(this, file);
-    }
-
-    public void DeleteUser(User x,String filename) throws IOException
-    {
-        File read_file = new File("/home/cybereagle3-1/AndroidStudioProjects/FirstApp/midsem_project/src/main/java/filestuff/" + filename);
-        new FileIO().DeleteUsersfromFile(x,read_file);
     }
 }