+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 5.11.17
 
-### commit
+### commit c8fa572b9cd996fb6f7dbb5f5ef178347be1c275
 
 #### Cop admin
 * created class CopAdmin extends Admin